requirements.txt
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.10
pytest==8.3.4
python-dotenv==1.0.1
PyJWT~=2.10.1
fastapi~=0.115.6
passlib~=1.7.4
pydantic~=2.10.4
asyncpg~=0.30.0
requests~=2.32.3
redis~=5.3.0b4